Tool to predict uncertainty: changed float16 to float32

+++ tools/predict_uncertainty.py
@@ -76,7 +76,7 @@ def main(args):
         Image.fromarray(heatmap).save(file_name)
         if args.store_raw_uncertainty:
             new_file_name = file_name + '_raw_uncertainty.npy'
-            np.save(new_file_name, uncertainty.astype(np.float16))
+            np.save(new_file_name, uncertainty)
         if args.store_prediction:
             new_file_name = file_name + '_prediction.png'
             Image.fromarray(pred.astype(np.uint8)).save(new_file_name)
